+++ src/secrets/mod.rs
@@ -8,6 +8,8 @@ pub mod kv1;
 pub mod kv2;
 #[cfg(feature = "pki")]
 pub mod pki;
+#[cfg(feature = "ssh")]
+pub mod ssh;
 #[cfg(feature = "totp")]
 pub mod totp;
 #[cfg(feature = "transit")]